+++ example/index.dart
@@ -4,8 +4,8 @@ import 'common.dart';
 void main() {
   final tokenInput = querySelector('#token') as InputElement;
   tokenInput.value = github.auth!.token ?? '';
-  window.localStorage['GITHUB_TOKEN'] = tokenInput.value!;
+  window.sessionStorage['GITHUB_TOKEN'] = tokenInput.value!;
   tokenInput.onKeyUp.listen((_) {
-    window.localStorage['GITHUB_TOKEN'] = tokenInput.value!;
+    window.sessionStorage['GITHUB_TOKEN'] = tokenInput.value!;
   });
 }