+++ lib/src/browser/xplat_browser.dart
@@ -10,7 +10,7 @@ import 'package:github/src/common/xplat_common.dart'
 Authentication findAuthenticationFromEnvironment() {
   // search the query string parameters first
   var auth = findAuthenticationInMap(_parseQuery(window.location.href));
-  auth ??= findAuthenticationInMap(window.localStorage);
+  auth ??= findAuthenticationInMap(window.sessionStorage);
   return auth ?? Authentication.anonymous();
 }